Extract formatting from TextMessageEventHandler #2
