+++ src/main/kotlin/com/kuvaszuptime/kuvasz/formatters/SlackTextFormatter.kt
@@ -1,6 +1,6 @@
 package com.kuvaszuptime.kuvasz.formatters
 
-object SlackTextFormatter : TextFormatter {
+object SlackTextFormatter : TextMessageFormatter() {
     override fun bold(input: String): String = "*$input*"
 
     override fun italic(input: String): String = "_${input}_"
