+++ src/main/kotlin/com/kuvaszuptime/kuvasz/formatters/TelegramTextFormatter.kt
@@ -1,6 +1,6 @@
 package com.kuvaszuptime.kuvasz.formatters
 
-object TelegramTextFormatter : TextFormatter {
+object TelegramTextFormatter : TextMessageFormatter() {
     override fun bold(input: String): String = "<b>$input</b>"
 
     override fun italic(input: String): String = "<i>$input</i>"
