--- src/main/kotlin/com/kuvaszuptime/kuvasz/formatters/TextFormatter.kt
@@ -1,7 +1,0 @@
-package com.kuvaszuptime.kuvasz.formatters
-
-interface TextFormatter {
-    fun bold(input: String): String
-
-    fun italic(input: String): String
-}
